mds: Store symlink target on the first data object
Problem:
The MDS only stores the symlink inode's backtrace
information in the data pool. During disaster
recovery of the metadata pool by scanning data
pool, the symlinks are recreated as regular files.

Solution:
This patch stores the symlink target on the first
data object as an xattr for recovery.

MDS option:
The mds option 'mds_symlink_recovery' is introduced
which is disabled by default. Enabling the option
stores the symlink target.

Signed-off-by: Kotresh HR <khiremat@redhat.com>
Fixes: https://tracker.ceph.com/issues/46166