qa/test_backtrace: Validate symlink xattr is stored
Validate symlink xattr is stored on first data
object of the symlink along with backtrace
if 'mds_symlink_recovery' option is enabled
and vice-versa.

Also add 'string_wrapper' class to decode bufferlist
to string. This helps 'ceph-dencoder' tool to decode
the symlink target stored which is used in tests to
validate.

Signed-off-by: Kotresh HR <khiremat@redhat.com>
Fixes: https://tracker.ceph.com/issues/46166