qa: reduce fs:workload use of random selection
It's more appropriate to use --subset to reduce the scheduling size. It
was previously laid out this way because we wanted to link to the common
`qa/cephfs/mount` directory so that ceph-fuse mounts are not needlessly
multiplied. We should just organize it correctly so that is not an
issue.

Signed-off-by: Patrick Donnelly <pdonnell@redhat.com>