rbd-mirror/journal: add image watcher to journal replayer
With multipeer, the rbd-mirror can't detect anymore that the
primary has been removed anymore with the ImageMap since it
would contain more peer than the primary remote and the local one.

This commit add an image watcher to the journal replayer like it has
been implemented in the snapshot mode. This allows to restart the
Replayer on any change made to the remote image including deletion and
renaming.

Signed-off-by: Arthur Outhenin-Chalandre <arthur.outhenin-chalandre@cern.ch>